코드 리뷰 반영

+++ app/src/main/java/com/example/bob_friend_android/viewmodel/DetailBoardViewModel.kt
@@ -18,19 +18,21 @@ class DetailBoardViewModel(application: Application): AndroidViewModel(applicati
     fun delete(context: Context, id : Int) {
         val token = App.prefs.getString("token", "no token")
 
-        RetrofitBuilder.api.deleteRecruitmens(token, id).enqueue(object : Callback<Void> {
-            override fun onResponse(call: Call<Void>, response: Response<Void>) {
-                Toast.makeText(context, "삭제되었습니다.", Toast.LENGTH_SHORT).show()
-                val intent = Intent(context, MainActivity::class.java)
-                intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP)
-                context.startActivity(intent)
-            }
+        if (token != null) {
+            RetrofitBuilder.api.deleteRecruitmens(token, id).enqueue(object : Callback<Void> {
+                override fun onResponse(call: Call<Void>, response: Response<Void>) {
+                    Toast.makeText(context, "삭제되었습니다.", Toast.LENGTH_SHORT).show()
+                    val intent = Intent(context, MainActivity::class.java)
+                    intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP)
+                    context.startActivity(intent)
+                }
 
-            override fun onFailure(call: Call<Void>, t: Throwable) {
-                Toast.makeText(context, "서버에 연결이 되지 않았습니다. 다시 시도해주세요!", Toast.LENGTH_SHORT).show()
-                Log.e("DetailActivity!!!", t.message.toString())
-            }
+                override fun onFailure(call: Call<Void>, t: Throwable) {
+                    Toast.makeText(context, "서버에 연결이 되지 않았습니다. 다시 시도해주세요!", Toast.LENGTH_SHORT).show()
+                    Log.e("DetailActivity!!!", t.message.toString())
+                }
 
-        })
+            })
+        }
     }
 }
